Add periodic checks to ensure sfctl version is compatible with cluster version (#123)
Every 24 hours or more (call made on sfctl command invocation, rather than on a timer), or on failure and on cluster connect, a check will be made against the cluster to verify that the cluster version is correct. This applies only to runtime 6.4 and sfctl version 7.0.0 and going forward. Caller will get a warning message on cluster version mismatch. Help commands are not checked.

Reviewed by Jeffrey, but approval was dismissed by some small changes.