Add batched telemetry (#166)
Improve upon #152 to make telemetry send batched telemetry in order to improve performance.

Telemetry will write to a file and hold on to telemetry to send in chunks. Current value is 50 items per batch. There are limited retries, so that if there is something such as no internet connection, we retry intermittently